up_board: bump toolchain and Linux kernel
This bumps the toolchain here even though it's not necessary. This lets
us match toolchain configurations with what's in use for production
images for ease of maintenance. The toolchain bump pulls in gcc 10 and
requires a Linux kernel bump to a later 4.14 patch release. Note that
this kernel boots with previous Linux kernel even though the prod
firmware wouldn't. I don't know the reason why, but it seems better to
match nonetheless.